move gradle into java-simple-application
