+++ generators/base-core/generator.ts
@@ -48,7 +48,7 @@ import { CRLF, LF, type Logger, hasCrlf, normalizeLineEndings, removeFieldsWithN
 import baseCommand from '../base/command.ts';
 import { dockerPlaceholderGenerator } from '../docker/utils.ts';
 import { GENERATOR_JHIPSTER } from '../generator-constants.ts';
-import { getGradleLibsVersionsProperties } from '../gradle/support/dependabot-gradle.ts';
+import { getGradleLibsVersionsProperties } from '../java-simple-application/generators/gradle/support/dependabot-gradle.ts';
 import type GeneratorsByNamespace from '../types.ts';
 import type { GeneratorsWithBootstrap } from '../types.ts';
