+++ generators/generate-blueprint/internal/lookup-namespaces.spec.ts
@@ -55,7 +55,6 @@ describe('lookupGeneratorsNamespaces', () => {
   "export-jdl",
   "generate-blueprint",
   "git",
-  "gradle",
   "heroku",
   "info",
   "init",
@@ -65,6 +64,7 @@ describe('lookupGeneratorsNamespaces', () => {
   "java-simple-application:build-tool",
   "java-simple-application:code-quality",
   "java-simple-application:graalvm",
+  "java-simple-application:gradle",
   "java-simple-application:jib",
   "java-simple-application:maven",
   "java-simple-application:openapi-generator",
