+++ generators/java-simple-application/generators/code-quality/generator.spec.ts
@@ -78,7 +78,7 @@ describe(`generator - ${generator}`, () => {
     it('should compose with generators', () => {
       expect(result.getComposedGenerators()).toMatchInlineSnapshot(`
 [
-  "jhipster:gradle",
+  "jhipster:java-simple-application:gradle",
 ]
 `);
     });
