+++ generators/java-simple-application/generators/gradle/command.ts
@@ -16,7 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-import type { JHipsterCommandDefinition } from '../../lib/command/index.ts';
+import type { JHipsterCommandDefinition } from '../../../../lib/command/index.ts';
 
 const command = {
   configs: {
