+++ generators/java-simple-application/generators/gradle/constants.ts
