+++ generators/java-simple-application/generators/gradle/generator.ts
@@ -19,10 +19,10 @@
 
 import assert from 'node:assert/strict';
 
-import { mutateData } from '../../lib/utils/index.ts';
-import { QUEUES } from '../base-core/priorities.ts';
-import BaseSimpleApplicationGenerator from '../base-simple-application/index.ts';
-import { GRADLE_BUILD_SRC_DIR } from '../generator-constants.ts';
+import { mutateData } from '../../../../lib/utils/index.ts';
+import { QUEUES } from '../../../base-core/priorities.ts';
+import BaseSimpleApplicationGenerator from '../../../base-simple-application/index.ts';
+import { GRADLE_BUILD_SRC_DIR } from '../../../generator-constants.ts';
 
 import { GRADLE } from './constants.ts';
 import files from './files.ts';
