+++ generators/java-simple-application/generators/gradle/support/dependabot-gradle.ts
