+++ generators/java-simple-application/generators/gradle/support/index.ts
