+++ generators/java-simple-application/generators/gradle/templates/.prettierignore.jhi.gradle.ejs
@@ -17,7 +17,7 @@
  limitations under the License.
 -%>
 <&_ if (!fragment.section) { -&>
-# Generated by jhipster:gradle
+# Generated by jhipster:java-simple-application:gradle
 build
 gradle
 .gradle
