+++ generators/java-simple-application/generators/gradle/templates/buildSrc/build.gradle.ejs
