+++ generators/java-simple-application/generators/gradle/templates/buildSrc/gradle/libs.versions.toml.ejs
