+++ generators/java-simple-application/generators/gradle/templates/gradle/libs.versions.toml.ejs
