+++ generators/java-simple-application/generators/gradle/templates/gradlew.bat
