+++ generators/java-simple-application/types.d.ts
@@ -27,7 +27,6 @@ import type {
   Options as BaseSimpleApplicationOptions,
   Source as BaseSimpleApplicationSource,
 } from '../base-simple-application/types.ts';
-import type { GradleNeedleOptions, Source as GradleSource } from '../gradle/types.ts';
 import type { JavaAnnotation } from '../java/support/add-java-annotation.ts';
 
 import type { JavaSimpleApplicationAddedApplicationProperties } from './application.ts';
@@ -37,6 +36,7 @@ import type {
   Config as BuildToolConfig,
   Options as BuildToolOptions,
 } from './generators/build-tool/types.ts';
+import type { GradleNeedleOptions, Source as GradleSource } from './generators/gradle/types.ts';
 import type { MavenDefinition, Source as MavenSource } from './generators/maven/types.ts';
 
 type Command = HandleCommandTypes<typeof command>;
